clk: si5351: In probe, ensure the device is ready
During power up the device copies the content of the NVM into RAM and
performs a system initialization. The device is not operational until
initialization is complete. It is not recommended to read or write
registers in RAM through the I2C interface until initialization is
complete [AN619].

Referenced document:
[AN619] Manually Generating an Si5351 Register Map, Rev. 0.4, page 14.

Signed-off-by: Sergej Sawazki <ce3a@gmx.de>