ASoC: core: Add lsb_first option to multi register control
The multiple codec register callbacks provide access to the value of a control
that can span multple codec registers which together form a single value in a
MSB/LSB manner. However, the assumption that the first register (regbase) is
the MSB part of the value is not true for all codecs. In the WM8741 codec, for
example, the first register represents the LSB part of the value.

This patch adds the lsb_first option to multiple codec register callbacks.
 - lsb_first=0 means regbase is the MSB part of the data.
 - lsb_first=1 means regbase is the LSB part of the data.